backend/pyproject.toml
[project]
name = "deepfake-detector-backend"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-multipart",
    "pillow",
    "torch==2.5.1+cu121",
    "torchvision==0.20.1+cu121",
]
requires-python = ">=3.12,<3.13"

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }